Refactor: Move proxy and CA handling to shared http utility
